Search Algorithms
The task in this program is to create a search algorithm to help a robot find its way to a destination. The robot exists in a grid world named “map” of size MAP_WIDTH by MAP_HEIGHT, and it can move in all 4 directions (diagonals not allowed) through empty space cells only by steps of exactly 1 cell distance.