Informed Greedy Search
The task in this program is to use greedy hill-climbing search to find minima and solutions to the 10-queens problem. This problem tries to find a position for all 10 queens on the board where they cannot attack each other. Reminder, an attack is whenever a queen shares a row, column, or diagonal with another queen. Here we will provide an initial configuration of the board for the 10-queens problem, and you are to modify that board with the local minima reached from that starting board. Here, a single step in the hill-climb is the movement of any one of the queens.