A Star Search
The task in this program is to create a search algorithm to help a robot find its way to a destination using A star search and the manhattan distance heuristic. The robot exists in a grid world named “map” of size MAP_WIDTH by MAP_HEIGHT, and it can move in all 4 directions (diagonals not allowed) through empty space cells only by steps of exactly 1 cell distance. The starting location of the robot is marked in the map with the number “2” and the goal with the number “3”. The other two values you can find on the map are “1” for walls and “0” for empty space.