predict result of a specific tic-tac-toe
The task in this program is to create a TIC-TAC-TOE solver capable of predicting the result of a specific game when a board is provided. The solver has to use min-max to predict the outcome as if both players were playing optimally. Your program will receive an array of 9 (from 0 to 8) integers representing the squares of the tic tac toe board and a separate number indicating which player moves next (1 for X, 2 for O).